# openapi.yaml
openapi: "3.0.3"
info:
  title: CampusXchange API
  version: "1.0.0"
  description: |
    Student peer-to-peer marketplace REST API.

    **Base URL:** `http://localhost:8080`
    **Swagger UI:** `http://localhost:8080/swagger-ui.html`

    ## Authentication
    Protected endpoints require two headers:
    ```
    Authorization: Bearer <access_token>
    X-User-Id: <userId>
    ```
    Both values are returned by `/api/auth/login`, `/api/auth/register`, and `/api/auth/google`.

    ## Real-time Chat (WebSocket)
    Connect via SockJS + STOMP to `http://localhost:8080/ws`.
    - Publish to `/app/chat/send` (include `X-User-Id` as STOMP header)
    - Subscribe to `/topic/chat/{min_uid}_{max_uid}` where the UIDs are sorted ascending
servers:
  - url: http://localhost:8080
    description: Local development
security: []   # overridden per-operation where required
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: student@college.edu
        password:
          type: string
          minLength: 6
          example: secret123
    RegisterRequest:
      type: object
      required: [username, email, firstName, lastName, password, college]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        email:
          type: string
          format: email
        firstName:
          type: string
          minLength: 2
        lastName:
          type: string
          minLength: 2
        password:
          type: string
          minLength: 8
        confirmPassword:
          type: string
          description: Validated client-side only; backend ignores this field
        college:
          type: string
          minLength: 2
        phoneNumber:
          type: string
          maxLength: 20
        studentId:
          type: string
          maxLength: 50
    GoogleAuthRequest:
      type: object
      required: [credential]
      properties:
        credential:
          type: string
          description: Google ID token returned by the Google Sign-In flow
    AuthResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
        fullName:
          type: string
        profilePhotoUrl:
          type: string
          nullable: true
        role:
          $ref: "#/components/schemas/UserRole"
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          description: Token lifetime in milliseconds
          example: 86400000
        emailVerified:
          type: boolean
        studentVerified:
          type: boolean

    # ── Users ─────────────────────────────────────────────────────────────
...
    CreateMessageRequest:
      type: object
      required: [recipientId, content]
      properties:
        recipientId:
          type: integer
          format: int64
        content:
          type: string
          minLength: 1
          maxLength: 5000
        imageUrl:
          type: string
          format: uri
          nullable: true
    MessageDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        senderId:
          type: integer
          format: int64
        senderName:
          type: string
        recipientId:
          type: integer
          format: int64
        recipientName:
          type: string
        content:
          type: string
        imageUrl:
          type: string
          format: uri
          nullable: true
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
        readAt:
          type: string
          format: date-time
          nullable: true
    ConversationSummary:
      type: object
      description: Returned by GET /api/messages/conversations
      properties:
        userId:
          type: integer
          format: int64
          description: Conversation partner's user ID
        username:
          type: string
        fullName:
          type: string
        profilePhotoUrl:
          type: string
          nullable: true
        unreadCount:
          type: integer

    # ── Enums ─────────────────────────────────────────────────────────────
    UserRole:
      type: string
      enum: [STUDENT, ADMIN]
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
          example: EMAIL_EXISTS
        message:
          type: string
        path:
          type: string
    ValidationErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        error:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Validation failed
        errors:
          type: object
          additionalProperties:
            type: string
          example:
            email: "Email must be valid"
            password: "Password must be between 8 and 50 characters"
    Page:
      type: object
      description: Spring Data pagination wrapper
      properties:
        content:
          type: array
          items: {}
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
          description: Current page (0-indexed)
        first:
          type: boolean
        last:
          type: boolean
    MessageResponse:
      type: object
      properties:
        message:
          type: string

# ── Paths ──────────────────────────────────────────────────────────────────
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new student account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "409":
          description: Email or username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/google:
    post:
      tags: [Auth]
      summary: Login or register via Google OAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleAuthRequest"
      responses:
        "200":
          description: OK — existing user logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "201":
          description: Created — new account auto-registered from Google profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid or expired Google token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password reset link
      description: Always returns 200 regardless of whether the email exists (prevents user enumeration).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token using a refresh token
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
            example: "Bearer <refresh_token>"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid or expired refresh token
...
  /api/messages/conversations:
    get:
      tags: [Messages]
      summary: List all conversation partners with unread count
      security:
        - BearerAuth: []
      parameters:
        - in: header
          name: X-User-Id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConversationSummary"
        "401":
          description: Unauthenticated
  /api/messages/conversation/{partnerId}:
    get:
      tags: [Messages]
      summary: Get paginated message history with one partner
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: partnerId
          required: true
          schema:
            type: integer
            format: int64
        - in: header
          name: X-User-Id
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      content:
                        type: array
                        items:
                          $ref: "#/components/schemas/MessageDTO"
        "401":
          description: Unauthenticated
  /api/messages/send:
    post:
      tags: [Messages]
      summary: Send a message via REST (WebSocket preferred for real-time)
      security:
        - BearerAuth: []
      parameters:
        - in: header
          name: X-User-Id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageDTO"
        "401":
          description: Unauthenticated
  /api/messages/unread:
    get:
      tags: [Messages]
      summary: Get all unread messages for the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - in: header
          name: X-User-Id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MessageDTO"
        "401":
          description: Unauthenticated
  /api/messages/{id}/read:
    put:
      tags: [Messages]
      summary: Mark a specific message as read
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageDTO"
  /api/messages/{id}:
    delete:
      tags: [Messages]
      summary: Delete a message
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...