# src/main/java/com/campusxchange/entity/UserRole.java
package com.campusxchange.entity;

/**
 * User Role Enum - Defines different roles in the CampusXchange system
 */
public enum UserRole {
    STUDENT,    // Regular student user
    SELLER,     // Student who sells products
    ADMIN,      // Platform administrator
    MODERATOR   // Content moderator
}